third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/historical/popup-same-site-with-cross-origin.https-expected.txt

This is a testharness.js-based test.
PASS same-site document opening popup to https://not-web-platform.test:8444 with COOP: ""
PASS same-site document opening popup to https://not-web-platform.test:8444 with COOP: "unsafe-none"
PASS same-site document opening popup to https://not-web-platform.test:8444 with COOP: "jibberish"
PASS same-site document opening popup to https://not-web-platform.test:8444 with COOP: "same-site"
PASS same-site document opening popup to https://not-web-platform.test:8444 with COOP: "same-site unsafe-allow-outgoing"
PASS same-site document opening popup to https://not-web-platform.test:8444 with COOP: "same-origin unsafe-allow-outgoing"
FAIL same-site document opening popup to https://not-web-platform.test:8444 with COOP: "same-origin" assert_equals: Openee browsing context closed expected true but got false
FAIL same-site document opening popup to https://not-web-platform.test:8444 with COOP: "same-origin-allow-popups" assert_equals: Openee browsing context closed expected true but got false
Harness: the test ran to completion.
